if no group results are show for user after interest selection, they create a new group
